9 personal profile create personal profile (user) (#22)
* Feat: Add two personal profile pages, one for admin and one for user

Initial setup.
Removed an irrelevant comment.
Added a space just to make things easier to read.

* Fix: Bringin the ID along.  You can now see the user information

Still to fix: Rest of the user information.
Cleaning up

* Fix: User profile working with all variables.

To update: Complete admin page.

* Fix: Remove adminPage - it will be made in a different issue. Update language from swedish to english

Realised we possibly have to add another endpoint for the adminpage. Decided to break it out and make a separate issue for it. Therefor removed it from this branch.